Fixed problem with pub user

+++ resources/views/partials/auctionFull.blade.php
@@ -13,7 +13,7 @@
 
 
 
-    @if ( (Auth::check() && Auth::id() == $auction->seller_id ) or Auth::user()->is_admin)
+    @if ( (Auth::check() && Auth::id() == $auction->seller_id ) or (Auth::check() && Auth::user()->is_admin))
       <a href="/auctions/{{ $auction->auction_id }}/delete" class="btn btn-secondary">Delete</a>
       <a href="/auctions/{{ $auction->auction_id }}/edit" class="btn btn-secondary">Edit</a>
     @else
